split modules more logically
